dht22.py first version
This is the first version of the dht22.py script. There are a lot of errors, because the original c code on which this version is bases has also errors. Sometimes it isn´t possible to get any data from the sensor. Maybe it will be fixed in other versions